Added projects
These are assignments I did in my last class at UHD, Predictive Analytics. Assignment 2 is was a showcase of various analysis methods on a single dataset written in R using RStudio. Assignment 3 was an attempt to use hyperparameter tuning to train a CNN on a small dataset that had been expanded using augmentation. I might update these in the future to match my current programming style.